--- back_end/api_route.py
@@ -1,12 +1,0 @@
-from flask import Flask, request, jsonify, redirect ,render_template
-app = Flask(__name__)
-
-
-    
-
-
-
-
-
-if __name__ == '__main__':
-    app.run()
