+++ templates/index.html
@@ -8,8 +8,8 @@
 	<!-- Icon library -->
 	<script src="https://kit.fontawesome.com/202330189a.js" crossorigin="anonymous"></script>
 	<!-- .css links -->
-	<link rel="stylesheet" type="text/css" href="index.css">
-    <link rel="stylesheet" type="text/css" href="/css/style.css">
+	<link rel="stylesheet" type="text/css" href="static/css/index.css">
+    <link rel="stylesheet" type="text/css" href="static/css/style.css">
     <!-- Tab text & Icon -->
     <link rel="icon" href="/icon/home.png">
     <title>E.I. - Home</title>
@@ -47,7 +47,7 @@
                 <td>Als product owner ben je eigenlijk niet bezig...</td>
             </tr>
             <tr>
-                <td id="iets"></td>
+                <td id="vacature_id"></td>
                 <td>Product Owner</td>
                 <td>100.000</td>
                 <td>Als product owner ben je eigenlijk niet bezig...</td>
